requirements.txt
pyserial
tk
pyinstaller